✨ show transactions list

+++ backend/Models/Account.cs
@@ -3,7 +3,7 @@ namespace coinly.Models;
 public class Account
 {
     public int id { get; set; }
-    public double income { get; set; }
-    public double expense { get; set; }
-    public double balance { get; set; }
+    public float income { get; set; }
+    public float expense { get; set; }
+    public float balance { get; set; }
 }
